requirements.txt

python-nmap
pymetasploit3
fpdf
markdown
networkx
matplotlib
sqlite3
